move calendly access to the course associated

+++ server/routes/api/accessToken.js
@@ -1,46 +1,51 @@
+/* eslint-disable no-unused-vars */
 const router = require('express').Router();
 const { authorizeToken } = require('../../utils/auth');
-const { AccessToken, Tutor, Calendly } = require('../../models');
+const {
+  AccessToken, Tutor, Calendly, Course,
+} = require('../../models');
 const { encryptToken } = require('../../utils/encryption');
+const { getTutorById } = require('../../utils/helpers');
 
-router.post('/', authorizeToken, async (req, res) => {
+router.post('/:courseId', authorizeToken, async (req, res) => {
   try {
     //  get tutor document
-    const tutor = await Tutor.findById(req.tutor._id);
+    const { tutor } = await getTutorById(req.tutor._id);
     if (!tutor) return res.status(401).json('unauthorized');
     // compare request password to tutor saved password
     if (!await tutor.isCorrectPassword(req.body.password)) return res.status(401).json('unauthorized');
 
     //  encrypt token with current password
     const encryptedToken = encryptToken(req.body.token, req.body.password);
-    const existingKeyId = tutor.accessTokens.calendly;
-    // if a token already exists
-    if (existingKeyId) {
+    const { calendly: { accessToken } } = await Course.findById(req.params.courseId);
+
+    // if no token exists
+    if (!accessToken) {
       try {
-        // update existing document with new token
-        await AccessToken.findByIdAndUpdate(existingKeyId, { token: encryptedToken });
-        return res.json('updated existing token');
+      // create db entry
+        const { _id } = await AccessToken.create({ token: encryptedToken });
+        if (!_id) return res.status(500).json('failed to create token');
+
+        // add new document id to tutor
+        const updated = await Course.findByIdAndUpdate(
+          req.params.courseId,
+          { calendly: { accessToken: _id, data: null } },
+        );
+        if (!updated) return res.status(500).json('failed to update user');
+
+        // send new token id back to client to update local state
+        return res.json({ _id });
       } catch (error) {
         console.error(error.message);
         return res.status(500).json({ location: 2, message: error.message });
       }
     }
 
-    // if no token exists
+    // if a token already exists
     try {
-      // create db entry
-      const { _id } = await AccessToken.create({ token: encryptedToken });
-      if (!_id) return res.status(500).json('failed to create token:0');
-
-      // add token document id to tutor
-      const updated = await Tutor.findByIdAndUpdate(
-        req.tutor._id,
-        { accessTokens: { calendly: _id } },
-      );
-      if (!updated) return res.status(500).json('failed to update user');
-
-      // send new token id back to client to update local state
-      return res.json({ _id });
+      // update existing document with new token
+      await AccessToken.findByIdAndUpdate(accessToken, { token: encryptedToken });
+      return res.json({ _id: accessToken });
     } catch (error) {
       console.error(error.message);
       return res.status(500).json({ location: 3, message: error.message });
@@ -51,15 +56,15 @@ router.post('/', authorizeToken, async (req, res) => {
   }
 });
 
-router.delete('/', authorizeToken, async ({ tutor, body: { tokenId } }, res) => {
+router.delete('/:tokenId', authorizeToken, async ({
+  params: { tokenId },
+  body: { courseId },
+}, res) => {
   try {
-    const { calendly } = await Tutor.findByIdAndUpdate(
-      tutor._id,
-      { calendly: null, accessTokens: { calendly: null } },
-    );
-
-    await Calendly.findByIdAndDelete(calendly);
-    await AccessToken.findByIdAndDelete(tokenId);
+    const { calendly: { accessToken, data } } = await Course.findById(courseId);
+    if (accessToken) await AccessToken.findByIdAndDelete(tokenId);
+    if (data) await Calendly.findByIdAndDelete(data);
+    await Course.findByIdAndUpdate(courseId, { calendly: { accessToken: null, data: null } });
 
     return res.json('token deleted');
   } catch ({ message }) {
